Update TokenFormat import paths
