+++ packages/tokens-studio-for-figma/src/plugin/store.ts
@@ -1,5 +1,3 @@
-type TokenFormatOptions = 'dtcg' | 'legacy';
-
 type State = {
   inspectDeep: boolean;
   shouldSendSelectionValues: boolean;
@@ -9,31 +7,5 @@ const store: State = {
   inspectDeep: false,
   shouldSendSelectionValues: false,
 };
-
-export class TokenFormatStoreClass {
-  public tokenValueKey: 'value' | '$value';
-
-  public tokenTypeKey: 'type' | '$type';
-
-  public tokenDescriptionKey: 'description' | '$description';
-
-  public format: 'dtcg' | 'legacy';
-
-  constructor() {
-    this.tokenValueKey = 'value';
-    this.tokenTypeKey = 'type';
-    this.tokenDescriptionKey = 'description';
-    this.format = 'legacy';
-  }
-
-  public setFormat = (format: TokenFormatOptions) => {
-    this.tokenValueKey = format === 'dtcg' ? '$value' : 'value';
-    this.tokenTypeKey = format === 'dtcg' ? '$type' : 'type';
-    this.tokenDescriptionKey = format === 'dtcg' ? '$description' : 'description';
-  };
-}
-
-export const TokenFormat = new TokenFormatStoreClass();
-export const { setFormat } = TokenFormat;
 
 export default store;
