+++ packages/tokens-studio-for-figma/src/utils/is/isTokenGroupWithType.ts
@@ -1,4 +1,4 @@
-import { TokenFormat } from '@/plugin/store';
+import { TokenFormat } from '@/plugin/TokenFormatStoreClass';
 import { SingleToken } from '@/types/tokens';
 
 type SingleTokenValueObject = Pick<SingleToken, 'value'>;
