+++ packages/tokens-studio-for-figma/src/utils/stringifyTokens.ts
@@ -2,7 +2,7 @@ import set from 'set-value';
 import { appendTypeToToken } from '@/app/components/createTokenObj';
 import { AnyTokenList } from '@/types/tokens';
 import removeTokenId from './removeTokenId';
-import { TokenFormat } from '@/plugin/store';
+import { TokenFormat } from '@/plugin/TokenFormatStoreClass';
 
 export function getGroupTypeName(tokenName: string, groupLevel: number): string {
   if (tokenName.includes('.')) {
@@ -23,19 +23,21 @@ export default function stringifyTokens(
     const { name, ...tokenWithoutName } = removeTokenId(tokenWithType, !storeTokenIdInJsonEditor);
     if (tokenWithoutName.inheritTypeLevel) {
       const {
-        type, inheritTypeLevel, value, ...tokenWithoutTypeAndValue
+        type, inheritTypeLevel, description, value, ...tokenWithoutTypeAndValue
       } = tokenWithoutName;
       // set type of group level
       set(tokenObj, getGroupTypeName(token.name, inheritTypeLevel), tokenWithoutName.type);
       set(tokenObj, token.name, tokenWithoutTypeAndValue);
       set(tokenObj, `${token.name}.${TokenFormat.tokenValueKey}`, value);
+      set(tokenObj, `${token.name}.${TokenFormat.tokenDescriptionKey}`, description);
     } else {
       const {
-        type, value, ...tokenWithoutTypeAndValue
+        type, value, description, ...tokenWithoutTypeAndValue
       } = tokenWithoutName;
       set(tokenObj, token.name, tokenWithoutTypeAndValue);
       set(tokenObj, `${token.name}.${TokenFormat.tokenTypeKey}`, type);
       set(tokenObj, `${token.name}.${TokenFormat.tokenValueKey}`, value);
+      set(tokenObj, `${token.name}.${TokenFormat.tokenDescriptionKey}`, description);
     }
   });
